@44736
	Eliminate the distinction between nested object types and
	normal object types. Now any object can be used in a nested
	context. This change involved modifying the Session API to
	require an explicit store call after create is called in order
	to assign any top level objects to an object map. Minor
	modifications to JDO, CAP, and various test code that use the
	Session API directly was required. This change also removes
	support for the "nested" PDL keyword, and modifies the jdo pdl
	generator to not use the keyword. There should be three
	additional test failures caused by this change. Two are
	testing error messages when nested types are used as top level
	types or when a non nested type is used in a nested context.
	Neither of these behaviors are illegal anymore. The third
	additional failure is due to a change in the behavior of the
	session with respect to duplicate session keys. Previously
	this would be flagged as an error on create. Now it is treated
	as a violation and doesn't get flagged as an error until
	flushAll() or commit().

+++ archive/eng/cap/test/src/com/arsdigita/persistence/NodeTest.java
@@ -27,12 +27,12 @@ import org.apache.log4j.Logger;
  * NodeTest
  *
  * @author <a href="mailto:rhs@mit.edu">rhs@mit.edu</a>
- * @version $Revision: #3 $ $Date: 2004/09/01 $
+ * @version $Revision: #4 $ $Date: 2004/09/30 $
  */
 
 public abstract class NodeTest extends PersistenceTestCase {
 
-    public final static String versionId = "$Id: //eng/persistence/dev/cap/test/src/com/arsdigita/persistence/NodeTest.java#3 $ by $Author: dennis $, $DateTime: 2004/09/01 11:40:07 $";
+    public final static String versionId = "$Id: //eng/persistence/dev/cap/test/src/com/arsdigita/persistence/NodeTest.java#4 $ by $Author: rhs $, $DateTime: 2004/09/30 15:44:52 $";
 
     private static Logger s_log =
         Logger.getLogger(NodeTest.class.getName());
@@ -106,15 +106,18 @@ public abstract class NodeTest extends PersistenceTestCase {
           catch (PersistenceException e) {
           }*/
 
+        DataObject node2 = null;
+        node = null;
         try {
-            node = createNode(0, "Root");
-            node.save();
+            node2 = createNode(0, "Root");
+            node2.save();
             node = createNode(0, "Root");
             node.save();
             fail("Created two nodes with same id!");
 
         } catch (PersistenceException e) {
-            node.delete();
+            if (node2 != null) { node2.delete(); }
+            if (node != null) { node.delete(); }
         }
 
         try {
